TYP: move stub-only helper types to pandas-stubs/_stubs_only
Move the arithmetic aliases, ElementOpsMixin, Supports_Proto*
protocols, and T_INTERVAL_NP out of pandas.core.base into the
private pandas._stubs_only module so the public stub surface matches
the pandas runtime.
